Base Types

Abstract Base Types

AbstractFeedbackNet

AbstractFeedbackNet is the base type for networks that can handle feedback connections.

FeedbackNets.AbstractFeedbackNets.AbstractFeedbackNet — Type.

AbstractFeedbackNet

Abstract base type for networks that include handling for feedback.

Interface

Any subtype should support iteration (over its layers) in order for the generic method of this type to work.

source

Every type that inherits from AbstractFeedbackNet should support iteration over its layers. This is used to implement the splitnames and namesvalid functions in a generic manner.

FeedbackNets.AbstractFeedbackNets.splitnames — Method.

splitnames(net::AbstractFeedbackNet)

Return the names of all Splitters in net.

source

FeedbackNets.AbstractFeedbackNets.namesvalid — Method.

namesvalid(net::AbstractFeedbackNet)

Check if the input names of all Mergers in net have a corresponding Splitter and that no two Splitters have the same name.

source

AbstractMerger

AbstractMerger is the base type for layers that merge several streams (e.g., feedforward and feedback). Any type that inherits from it should implement a function to apply an instance of it to an input and a state dictionary from which to get the feedback streams.

FeedbackNets.AbstractMergers.AbstractMerger — Type.

AbstractMerger

Abstract base type for mergers.

Interface

Any subtype should support to combine a forward stream with other streams that can be accessed through a state dictionary via their Splitter name.

source