Network Types

Network Wrappers

Two types of networks are currently implemented: FeedbackChains and FeedbackTrees. For a comparison, see Controlling Information Flow: Chains vs Trees

FeedbackChains

FeedbackNets.FeedbackChains.FeedbackChain — Type.

FeedbackChain{T<:Tuple}

Tuple-like structure similar to a Flux.Chain with support for Splitters and Mergers.

source

FeedbackNets.FeedbackChains.FeedbackChain — Method.

(c::FeedbackChain)(h, x)

Apply a FeedbackChain to input x with hidden state h. h should take the form of a dictionary mapping Splitter names to states.

source

FeedbackTrees

FeedbackNets.FeedbackTrees.FeedbackTree — Method.

(c::FeedbackTree)(h, x)

Apply a FeedbackTree to input x with hidden state h. h should take the form of a dictionary mapping Splitter names to states.

source